move the dfu_confirm stuff to bm_dfu_generic/wrapper and remove all sys_cfg stuff from dfu! Next we just need to tackle the NVM parition stuff!
