+++ src/apps/bm_devkit/hello_world/CMakeLists.txt
@@ -105,7 +105,7 @@ set(LIB_FILES
     ${SRC_DIR}/lib/drivers/protected/protected_i2c.c
     ${SRC_DIR}/lib/drivers/protected/protected_spi.c
     ${SRC_DIR}/lib/drivers/abstract/abstract_spi.cpp
-    ${SRC_DIR}/lib/drivers/bm_dfu_wrapper.c
+    ${SRC_DIR}/lib/drivers/bm_dfu_wrapper.cpp
     ${SRC_DIR}/lib/drivers/bm_rtc_wrapper.c
     ${SRC_DIR}/lib/drivers/stm32_adc.c
     ${SRC_DIR}/lib/drivers/stm32_io.c
