+++ src/lib/bcmp/bm/bristlemouth.cpp
@@ -100,7 +100,7 @@ void bcl_init(NvmPartition *dfu_partition, cfg::Configuration *sys_cfg) {
       .sn = {'0', '1', '2', '3', '4', '5', '6', '7', '8', '9', 'a', 'b', 'c', 'd', 'e', 'f'},
   };
 
-  bcmp_init(dfu_partition, sys_cfg, device);
+  bcmp_init(dfu_partition, device);
   bcmp_cli_init();
 
   bm_middleware_init(&netif, BM_MIDDLEWARE_PORT);
